scripts/requirements.txt
mongoengine==0.8.7
pymongo==2.7.1
Flask==0.10.1
configparser==3.3.0r1
flask-restful==0.3.2
requests==2.2.1
geojson==1.2.1
Flask-OAuth==0.12
python-dateutil==1.5
pylint==1.4.4
selenium
astroid==1.3.8
pep8==1.6.2
jira
python-jenkins
